+++ packages/ccdi-server/src/params/filter.rs
@@ -93,6 +93,12 @@ pub struct Sample {
     #[param(required = false, nullable = false)]
     pub anatomical_site: Option<String>,
 
+    /// Matches any sample where the `library_selection_method` field matches the string
+    /// provided.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    #[param(required = false, nullable = false)]
+    pub library_selection_method: Option<String>,
+
     /// Matches any sample where the `library_strategy` field matches the string
     /// provided.
     #[serde(default, skip_serializing_if = "Option::is_none")]